requirements.txt
beautifulsoup4==4.10.0
certifi==2021.5.30
lxml==4.6.3
PyYAML==5.4.1
requests==2.26.0
soupsieve==2.2.1
urllib3==1.26.6
